* Added gendarme suppress attributes.
* Rebuilt FwBuildTasks.dll.
* Added disposals to DictionaryConfigurationTreeControl.Designer.cs.
* Adjust XhtmlDocViewTests to prevent and harden against Missing
Dispose.
* Prevent Missing Dispose in DictionaryConfigurationListenerTests.

Conflicts:
	.gitignore
	Build/FwBuildTasks.dll
	Build/Linux.targets
	Build/Src/FwBuildTasks/CollectTargets.cs
	Build/TestTimeoutValues.xml
	Build/Windows.targets
	Build/build
	Build/mkall.targets
	Makefile
	Src/Common/COMInterfaces/IcuWrappers.cs
	Src/Common/Controls/XMLViews/XMLViews.csproj
	Src/Common/CoreImpl/DirectoryFinder.cs
	Src/Common/FieldWorks/FieldWorks.cs
	Src/Common/FieldWorks/FieldWorks.csproj
	Src/Common/SimpleRootSite/SimpleRootSite.cs
	Src/FDO/DomainServices/WritingSystemServices.cs
	Src/FDO/FDOTests/PersistingLayerTests.BEPPortTests.cs
	Src/FwCoreDlgs/FwCoreDlgControls/ConfigSenseLayout.cs
	Src/Generic/Generic.vcxproj
	Src/LexText/LexTextControls/LexTextControls.csproj
	Src/LexText/LexTextDll/FlexStylesXmlAccessor.cs
	Src/LexText/Morphology/MGA/MGA.csproj
	Src/LexText/ParserUI/ParserUI.csproj
	Src/LexText/ParserUI/WebPageInteractor.cs
	Src/MasterVersionInfo.txt
	Src/Utilities/BasicUtils/BasicUtilsTests/NoOpKeyboardController.cs
	Src/Utilities/BasicUtils/SimpleLogger.cs
	Src/XCore/xCore.csproj
	Src/views/views.vcxproj
	Src/xWorks/Archiving/ReapRamp.cs
	Src/xWorks/DictionaryPublicationDecorator.cs
	Src/xWorks/xWorks.csproj
	Src/xWorks/xWorksStrings.resx
	Src/xWorks/xWorksTests/xWorksTests.csproj

+++ Build/Src/FwBuildTasks/TestTask.cs
@@ -1,3 +1,6 @@
+// Copyright (c) 2015 SIL International
+// This software is licensed under the LGPL, version 2.1 or later
+// (http://www.gnu.org/licenses/lgpl-2.1.html)
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -100,7 +103,20 @@ namespace FwBuildTasks
 				{
 					try
 					{
-						process.Kill();
+						bool testLooksToHaveFinished;
+						lock(LockObject)
+						{
+							testLooksToHaveFinished = !string.IsNullOrEmpty(m_TestLog.FindLast(line => line.Contains(GetTestsCompletedString())));
+						}
+						if(testLooksToHaveFinished)
+						{
+							Log.LogMessage(MessageImportance.Normal, "Tests for {0} timed out, but appear to have finished. Giving 2 seconds for log to be written.", TestProgramName);
+							Thread.Sleep(2000);
+							// If the tests completed, don't fail with a timeout; if missing Dispose calls caused NUnit to hang, that should be caught as a different failure.
+							fTimedOut = false;
+						}
+						if(!process.HasExited) // If the tests looked to have finished the process might have exited normally by now
+							process.Kill(); // This will set the exit code to -1 and the suite will be added as a failed suite below
 					}
 					catch
 					{
@@ -163,6 +179,8 @@ namespace FwBuildTasks
 			// test or if we get a timeout we want to fail the build.
 			return retVal;
 		}
+
+		protected abstract string GetTestsCompletedString();
 
 		/// <summary>
 		/// Starts the process and handles errors.
